requirements.txt
# S3 CSV Downloader Requirements

# AWS SDK for Python
boto3>=1.34.0

# Environment variable management
python-dotenv>=1.0.0

# Progress bar
tqdm>=4.66.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.0
